Mensaje de confirmación añadido a paginas

+++ app/dashboard/accounts/page.tsx
@@ -3,12 +3,15 @@ import { Loader2, PlusIcon } from "lucide-react";
 import { useNewAccount } from "@/features/accounts/hooks/use-new-account";
 import { useGetAccounts } from "@/features/accounts/api/use-get-accounts";
 import { useDeleteAccounts } from "@/features/accounts/api/use-delete-accounts";
+import { useConfirm } from "@/hooks/use-confirm";
 import { Button } from "@/components/ui/button";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { DataTable } from "@/components/data-table";
 import { columns } from "./columns";
 
 export default function AccountsPage() {
+
+  const [confirm, ConfirmDialog] = useConfirm("Are you sure?", "You are going to delete all your accounts.")
 
   const { onOpen } = useNewAccount()
   const accountsQuery = useGetAccounts()
@@ -35,6 +38,7 @@ export default function AccountsPage() {
 
   return (
     <div className="max-w-screen-2xl mx-auto">
+      <ConfirmDialog />
       <Card className="shadow-none">
         <CardHeader>
           <div className="flex flex-col lg:flex-row gap-y-1 lg:gap-y-0 items-center justify-between">
@@ -57,11 +61,15 @@ export default function AccountsPage() {
             data={accounts}
             filterKey="name"
             columns={columns}
-            onDelete={(rows, setRowsSelected) => {
-              const ids = rows.map((row) => row.original.id)
-              deleteAccounts({ ids }, {
-                onSuccess: () => setRowsSelected({})
-              })
+            onDelete={async (rows, setRowsSelected) => {
+              const ok = await confirm()
+
+              if (ok) {
+                const ids = rows.map((row) => row.original.id)
+                deleteAccounts({ ids }, {
+                  onSuccess: () => setRowsSelected({})
+                })
+              }
             }}
             disabled={isPending}
           />
